将stock中的redis操作和分析操作近一步拆分，分别放到utils/cache.py和utils/stock_analysis.py中
